+++ Vaajak.Application/Dto/Primitives/PaginationRequestDTO.cs
@@ -4,7 +4,7 @@ namespace Vaajak.Application.Dto.Primitives
 {
     public class PaginationRequestDTO
     {
-        public int Page { get; set; } = 1;
+        public int PageNumber { get; set; } = 1;
         [Range(1, 50)] public int PageSize { get; set; } = 10;
         public bool Paging { get; set; } = true;
     }
