+++ Vaajak.Application/Vaajak.Application.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.4" />
+    <PackageReference Include="X.PagedList" Version="10.0.3" />
   </ItemGroup>
 
   <ItemGroup>
