+++ Vaajak.Domain/Repositories/Vocabs/IVocabsRepository.cs
@@ -6,9 +6,9 @@ namespace Vaajak.Domain.Repositories.Vocabs
 {
     public interface IVocabsRepository
     {
-        Task<IList<Vocab>> GetAllAsync(int page, int pageSize);
+        Task<IEnumerable<Vocab>> GetAllAsync();
         Task<Vocab?> GetByIdAsync(Guid id);
         // Task<Vocab?> CreateVocab(CreateVocabDto createVocabDto);
-        Task<Vocab> CreateVocab(Vocab vocab);
+        Task<Vocab?> CreateVocab(Vocab vocab);
     }
 }
